# [Core] Improve detokenization performance for prefill