--- a/vllm/engine/llm_engine.py
+++ b/vllm/engine/llm_engine.py
@@ -1,5 +1,5 @@
 import time
-from typing import Dict, Iterable, List, Optional, Tuple, Type, Union
+from typing import Iterable, List, Optional, Tuple, Type, Union
 
 from transformers import PreTrainedTokenizer
 
@@ -15,12 +15,11 @@ from vllm.engine.ray_utils import initialize_ray_cluster
 from vllm.logger import init_logger
 from vllm.outputs import RequestOutput
 from vllm.sampling_params import SamplingParams
-from vllm.sequence import (Logprob, SamplerOutput, Sequence, SequenceGroup,
+from vllm.sequence import (SamplerOutput, Sequence, SequenceGroup,
                            SequenceGroupOutput, SequenceOutput, SequenceStatus)
-from vllm.transformers_utils.tokenizer import (detokenize_incrementally,
-                                               convert_prompt_ids_to_tokens)
 from vllm.transformers_utils.tokenizer_group import (BaseTokenizerGroup,
                                                      get_tokenizer_group)
+from vllm.transformers_utils.detokenizer import Detokenizer
 from vllm.utils import Counter
 
 logger = init_logger(__name__)
@@ -98,6 +97,7 @@ class LLMEngine:
         self._verify_args()
 
         self._init_tokenizer()
+        self.detokenizer = Detokenizer(self.tokenizer)
         self.seq_counter = Counter()
 
         self.model_executor = executor_class(model_config, cache_config,
@@ -150,7 +150,7 @@ class LLMEngine:
         raise RuntimeError("LLMEngine should not be pickled!")
 
     def get_tokenizer(self) -> "PreTrainedTokenizer":
-        return self.tokenizer.get_lora_tokenizer()
+        return self.tokenizer.get_lora_tokenizer(None)
 
     def get_tokenizer_for_seq(self,
                               sequence: Sequence) -> "PreTrainedTokenizer":
@@ -359,7 +359,7 @@ class LLMEngine:
         # Process prompt logprobs
         prompt_logprobs = outputs.prompt_logprobs
         if prompt_logprobs is not None:
-            seq_group.prompt_logprobs = self._decode_prompt_logprobs(
+            seq_group.prompt_logprobs = self.detokenizer.decode_prompt_logprobs(
                 seq_group, prompt_logprobs)
 
         # Process samples
@@ -403,7 +403,7 @@ class LLMEngine:
             child_seqs.append((parent, parent))
 
         for seq, _ in child_seqs:
-            self._decode_sequence(seq, seq_group.sampling_params)
+            self.detokenizer.decode_sequence(seq, seq_group.sampling_params)
             self._check_stop(seq, seq_group.sampling_params)
 
         # Non-beam search case
@@ -696,120 +696,6 @@ class LLMEngine:
             time_per_output_tokens=time_per_output_tokens,
             time_e2e_requests=time_e2e_requests,
         )
-
-    def _decode_logprobs(
-            self,
-            seq: Sequence,
-            prms: SamplingParams,
-            logprobs: Dict[int, Logprob],
-            all_input_ids: List[int],
-            generated_token_id_and_text: Optional[Tuple[int,
-                                                        str]] = None) -> None:
-        if not logprobs:
-            return
-        tokenizer = self.get_tokenizer_for_seq(seq)
-        previous_tokens = all_input_ids[:-1]
-        for token_id, sample_logprob in logprobs.items():
-            if (generated_token_id_and_text is not None
-                    and token_id == generated_token_id_and_text[0]):
-                sample_logprob.decoded_token = generated_token_id_and_text[1]
-                continue
-            if (sample_logprob.decoded_token is None and token_id != -1):
-                all_input_ids_with_logprob = previous_tokens + [token_id]
-                (_, new_text, _, _) = detokenize_incrementally(
-                    tokenizer=tokenizer,
-                    all_input_ids=all_input_ids_with_logprob,
-                    prev_tokens=seq.tokens,
-                    prefix_offset=seq.prefix_offset,
-                    read_offset=seq.read_offset,
-                    skip_special_tokens=prms.skip_special_tokens,
-                    spaces_between_special_tokens=prms.
-                    spaces_between_special_tokens,
-                )
-                sample_logprob.decoded_token = new_text
-
-    def _decode_prompt_logprobs(
-            self, seq_group: SequenceGroup,
-            prompt_logprobs: List[Dict[int,
-                                       Logprob]]) -> List[Dict[int, Logprob]]:
-        prms = seq_group.sampling_params
-        # We can pick any sequence for the prompt.
-        seq = next(iter(seq_group.seqs_dict.values()))
-        # Only prompt, without the generated token.
-        prompt_token_ids = seq.get_token_ids()[:-1]
-        tokenizer = self.get_tokenizer_for_seq(seq)
-        prefix_offset = 0
-        read_offset = 0
-        prev_tokens = None
-        first_non_none_index = 0
-        for i, prompt_logprobs_for_token in enumerate(prompt_logprobs):
-            if not prompt_logprobs_for_token:
-                first_non_none_index = i + 1
-                continue
-            for token_id, sample_logprob in prompt_logprobs_for_token.items():
-                if (sample_logprob.decoded_token is None and token_id != -1):
-                    prompt_token_ids_with_token = prompt_token_ids[:i] + [
-                        token_id
-                    ]
-                    (new_tokens, new_text, prefix_offset,
-                     read_offset) = detokenize_incrementally(
-                         tokenizer=tokenizer,
-                         all_input_ids=prompt_token_ids_with_token,
-                         prev_tokens=prev_tokens,
-                         prefix_offset=prefix_offset,
-                         read_offset=read_offset,
-                         skip_special_tokens=prms.skip_special_tokens,
-                         spaces_between_special_tokens=prms.
-                         spaces_between_special_tokens,
-                     )
-                    if prev_tokens is None:
-                        prev_tokens = new_tokens
-                    else:
-                        prev_tokens += new_tokens
-                    sample_logprob.decoded_token = new_text
-        return prompt_logprobs[first_non_none_index:]
-
-    def _decode_sequence(self, seq: Sequence, prms: SamplingParams) -> None:
-        """Decodes the new token for a sequence."""
-        all_input_ids = seq.get_token_ids()
-        tokenizer = self.get_tokenizer_for_seq(seq)
-
-        # Convert prompt token IDs to tokens if necessary.
-        # Do it here so that we don't have to repeat this
-        # computation for each logprob.
-        if seq.tokens is None:
-            (seq.tokens, seq.prefix_offset,
-             seq.read_offset) = convert_prompt_ids_to_tokens(
-                 tokenizer=tokenizer,
-                 prompt_ids=all_input_ids[:-1],
-                 skip_special_tokens=prms.skip_special_tokens,
-             )
-
-        (new_tokens, new_output_text, prefix_offset,
-         read_offset) = detokenize_incrementally(
-             tokenizer=tokenizer,
-             all_input_ids=all_input_ids,
-             prev_tokens=seq.tokens,
-             prefix_offset=seq.prefix_offset,
-             read_offset=seq.read_offset,
-             skip_special_tokens=prms.skip_special_tokens,
-             spaces_between_special_tokens=prms.spaces_between_special_tokens,
-         )
-
-        self._decode_logprobs(seq,
-                              prms,
-                              seq.output_logprobs[-1],
-                              all_input_ids,
-                              generated_token_id_and_text=(all_input_ids[-1],
-                                                           new_output_text))
-
-        if seq.tokens is None:
-            seq.tokens = new_tokens
-        else:
-            seq.tokens.extend(new_tokens)
-        seq.prefix_offset = prefix_offset
-        seq.read_offset = read_offset
-        seq.output_text += new_output_text
 
     def _check_stop(self, seq: Sequence,
                     sampling_params: SamplingParams) -> None: