@Yard1 2024-03-22
@njhill I like that idea, though we'd need to figure out a good design to share the tokenizer object across the instances. I think the default assumption may be that each `Detokenizer` has it's own HF tokenizer, but we'd like it to be shared