@njhill 2024-03-22
> I think the default assumption may be that each `Detokenizer` has it's own HF tokenizer,

@Yard1 I'm not sure I follow why that would be the case or why it would matter? I don't see the problem with multiple `Detokenizer` instances referencing the same tokenizer?