@Yard1 2024-03-22
@njhill Oh yeah from technical standpoint it's all clear, but from design standpoint multiple instances sharing the same object makes it hard to realize at a glance whether that object is shared or separate, which may lead to issues later ("As a new developer, I want to modify the tokenizer in this sequence for some reason, so I will just do that without realizing it's shared across all sequences")