@njhill 2024-03-22
Ah, makes sense! though I think it's not uncommon for such things to be shared. I.e. the field is just seen as a pointer to the tokenizer used by this detokenizer. In any case maybe a comment on the field making clear that it's shared would help with that?

And I don't mean to imply that this PR should necessarily be held up for this change, could always be done as a follow-on.