@Yard1 2024-03-22
@njhill definitely! I think it would be a good followup (even just breaking up some of the big sequence.py classes would be good)